made classes more organized
